pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "healthopsenv"
version = "1.0.0"
description = "An OpenEnv-compatible healthcare operations workflow simulation environment"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "graders*", "utils*", "server*"]
